refactor(coin-module-boilerplate): rename `common-logic` to `logic`
