+++ .changeset/few-avocados-cover.md
@@ -1,0 +1,5 @@
+---
+"@ledgerhq/coin-module-boilerplate": patch
+---
+
+refactor(coin-module-boilerplate): rename `common-logic` to `logic`
