+++ libs/coin-modules/coin-module-boilerplate/src/bridge/broadcast.test.ts
@@ -1,9 +1,10 @@
 import { patchOperationWithHash } from "@ledgerhq/coin-framework/operation";
 import { Account, BroadcastArg } from "@ledgerhq/types-live";
-jest.mock("@ledgerhq/coin-framework/operation");
-jest.mock("../common-logic");
-import { broadcast as broadcastLogic } from "../common-logic";
+import { broadcast as broadcastLogic } from "../logic";
 import { broadcast } from "./broadcast";
+
+jest.mock("@ledgerhq/coin-framework/operation");
+jest.mock("../logic");
 
 describe("broadcast", () => {
   let patchOperationSpy: jest.SpyInstance;
