+++ libs/coin-modules/coin-module-boilerplate/src/bridge/broadcast.ts
@@ -1,6 +1,6 @@
 import { patchOperationWithHash } from "@ledgerhq/coin-framework/operation";
 import { AccountBridge } from "@ledgerhq/types-live";
-import { broadcast as broadcastLogic } from "../common-logic";
+import { broadcast as broadcastLogic } from "../logic";
 import { Transaction } from "../types";
 
 export const broadcast: AccountBridge<Transaction>["broadcast"] = async ({
