+++ libs/coin-modules/coin-module-boilerplate/src/logic/account/getBalance.ts
