+++ libs/coin-modules/coin-module-boilerplate/src/logic/account/getNextSequence.ts
