+++ libs/coin-modules/coin-module-boilerplate/src/logic/common.ts
