+++ libs/coin-modules/coin-module-boilerplate/src/logic/history/lastBlock.ts
