+++ libs/coin-modules/coin-module-boilerplate/src/logic/index.ts
