+++ libs/coin-modules/coin-module-boilerplate/src/logic/transaction/craftTransaction.ts
