+++ libs/coin-modules/coin-module-boilerplate/src/logic/utils.ts
