+++ libs/coin-modules/coin-module-boilerplate/src/logic/validateAddress.ts
